Use git history to find PRs in a tag instead of time
Fixes #617

Before this change PRs were associated with tags by the merge date of
the PR, regardless of git branching history. In this change the oldest
tag in which the merge commit of each PR occurs in is determined by
using the Github compare API. That association is then used to sort PRs
into their respective release.

This change uses caching of commit history fetching and comparisons for
speedup.